Extract ordering stuff to a concern
